# docs/swagger.yaml
basePath: /api
definitions:
...
  GAMERS-BE_internal_contest_application_dto.ContestResponse:
    properties:
      auto_start:
        type: boolean
      contest_id:
        type: integer
      contest_status:
        $ref: '#/definitions/GAMERS-BE_internal_contest_domain.ContestStatus'
      contest_type:
        $ref: '#/definitions/GAMERS-BE_internal_contest_domain.ContestType'
      created_at:
        type: string
      description:
        type: string
      discord_guild_id:
        type: string
      discord_text_channel_id:
        type: string
      ended_at:
        type: string
      game_point_table_id:
        type: integer
      game_type:
        $ref: '#/definitions/GAMERS-BE_internal_game_domain.GameType'
      max_team_count:
        type: integer
      modified_at:
        type: string
      started_at:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      total_point:
        type: integer
      total_team_member:
        type: integer
    type: object
  GAMERS-BE_internal_contest_application_dto.CreateContestRequest:
    properties:
      auto_start:
        type: boolean
      contest_type:
        $ref: '#/definitions/GAMERS-BE_internal_contest_domain.ContestType'
      description:
        type: string
      discord_guild_id:
        type: string
      discord_text_channel_id:
        type: string
      ended_at:
        type: string
      game_point_table_id:
        type: integer
      game_type:
        $ref: '#/definitions/GAMERS-BE_internal_game_domain.GameType'
      max_team_count:
        type: integer
      started_at:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      total_point:
        type: integer
      total_team_member:
        type: integer
    required:
    - contest_type
    - title
    type: object
  GAMERS-BE_internal_contest_application_dto.DiscordLinkRequiredResponse:
    properties:
      message:
        type: string
      oauth_url:
        type: string
    type: object
  GAMERS-BE_internal_contest_application_dto.UpdateContestRequest:
    properties:
      auto_start:
        type: boolean
      contest_status:
        $ref: '#/definitions/GAMERS-BE_internal_contest_domain.ContestStatus'
      contest_type:
        $ref: '#/definitions/GAMERS-BE_internal_contest_domain.ContestType'
      description:
        type: string
      discord_guild_id:
        type: string
      discord_text_channel_id:
        type: string
      ended_at:
        type: string
      game_point_table_id:
        type: integer
      game_type:
        $ref: '#/definitions/GAMERS-BE_internal_game_domain.GameType'
      max_team_count:
        type: integer
      started_at:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      total_point:
        type: integer
      total_team_member:
        type: integer
    type: object
...
  GAMERS-BE_internal_contest_domain.ContestStatus:
    enum:
    - PENDING
    - ACTIVE
    - FINISHED
    - CANCELLED
    type: string
    x-enum-varnames:
    - ContestStatusPending
    - ContestStatusActive
    - ContestStatusFinished
    - ContestStatusCancelled
  GAMERS-BE_internal_contest_domain.ContestType:
    enum:
    - TOURNAMENT
    - LEAGUE
    - CASUAL
    type: string
    x-enum-varnames:
    - ContestTypeTournament
    - ContestTypeLeague
    - ContestTypeCasual
...
  GAMERS-BE_internal_game_domain.GameType:
    enum:
    - VALORANT
    - LOL
    type: string
    x-enum-varnames:
    - GameTypeValorant
    - GameTypeLOL
  GAMERS-BE_internal_global_common_dto.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  GAMERS-BE_internal_global_response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gamers.com
    name: API Support
  description: API Server for GAMERS platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GAMERS API
  version: "1.0"
paths:
...
  /api/contests:
    get:
      consumes:
      - application/json
      description: Get all contests with pagination, sorting, and title search support
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: 'Sort field: created_at, started_at, ended_at (default: created_at)'
        in: query
        name: sort_by
        type: string
      - description: 'Sort order: asc, desc (default: desc)'
        in: query
        name: order
        type: string
      - description: Search by contest title (partial match)
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_global_common_dto.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get all contests with pagination
      tags:
      - contests
    post:
      consumes:
      - application/json
      description: Create a new contest with contest details
      parameters:
      - description: Contest creation request
        in: body
        name: contest
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.CreateContestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.DiscordLinkRequiredResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new contest
      tags:
      - contests
...
  /api/contests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete contest by contest ID
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a contest
      tags:
      - contests
    get:
      consumes:
      - application/json
      description: Get contest details by contest ID
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      summary: Get a contest by ID
      tags:
      - contests
    patch:
      consumes:
      - application/json
      description: Update contest details by contest ID
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contest update request
        in: body
        name: contest
        required: true
        schema:
          $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.UpdateContestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Update a contest
      tags:
      - contests
...
  /api/contests/{id}/start:
    post:
      consumes:
      - application/json
      description: Start a contest and migrate accepted applications to config (Leader
        only)
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Start a contest
      tags:
      - contests
  /api/contests/{id}/stop:
    post:
      consumes:
      - application/json
      description: Stop an active contest and transition it to finished status (Leader
        only)
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_contest_application_dto.ContestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Stop a contest
      tags:
      - contests
...
  /api/contests/me:
    get:
      consumes:
      - application/json
      description: Get all contests that the authenticated user has joined with pagination,
        sorting, and filtering support
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: 'Sort field: created_at, started_at, ended_at, point, contest_status
          (default: created_at)'
        in: query
        name: sort_by
        type: string
      - description: 'Sort order: asc, desc (default: desc)'
        in: query
        name: order
        type: string
      - description: 'Filter by contest status: PENDING, ACTIVE, FINISHED, CANCELLED'
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/GAMERS-BE_internal_global_common_dto.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/GAMERS-BE_internal_global_response.Response'
      security:
      - BearerAuth: []
      summary: Get contests I have joined
      tags:
      - contests
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"